+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
 ## 3.0.2 - 26.04.2021 - Thomas Kurth
 
 - Add Translation UI which helps users to simply contribute to the project (Invoke-M365DocTranslationUI).
+- Extend Azure AD Documentation
 
 ## 3.0.1 - 26.04.2021 - Thomas Kurth